fix: sort creator courses by latest activity
