+++ src/api/tests/service/shifu/test_shifu_draft_list.py
@@ -4,7 +4,7 @@ from datetime import datetime
 from decimal import Decimal
 
 import flaskr.dao as dao
-from flaskr.service.shifu.models import DraftShifu
+from flaskr.service.shifu.models import DraftOutlineItem, DraftShifu
 from flaskr.service.shifu.shifu_draft_funcs import get_shifu_draft_list
 
 
@@ -89,4 +89,78 @@ def test_get_shifu_draft_list_sorts_by_updated_at_desc_then_id_desc(app):
         "draft-sort-newer",
         "draft-sort-same-time-a",
         "draft-sort-older",
+    ]
+
+
+def test_get_shifu_draft_list_prefers_latest_outline_activity(app):
+    owner_bid = "draft-list-activity-owner"
+    with app.app_context():
+        DraftOutlineItem.query.filter(
+            DraftOutlineItem.shifu_bid.in_(
+                ["draft-activity-course", "draft-activity-reference"]
+            )
+        ).delete(synchronize_session=False)
+        DraftShifu.query.filter(
+            DraftShifu.created_user_bid == owner_bid,
+            DraftShifu.shifu_bid.in_(
+                ["draft-activity-course", "draft-activity-reference"]
+            ),
+        ).delete(synchronize_session=False)
+
+        _seed_draft(
+            shifu_bid="draft-activity-course",
+            title="Outline Activity",
+            owner_bid=owner_bid,
+            created_at=datetime(2026, 5, 10, 10, 0, 0),
+            updated_at=datetime(2026, 5, 10, 10, 0, 0),
+        )
+        _seed_draft(
+            shifu_bid="draft-activity-reference",
+            title="Reference Course",
+            owner_bid=owner_bid,
+            created_at=datetime(2026, 5, 11, 10, 0, 0),
+            updated_at=datetime(2026, 5, 12, 10, 0, 0),
+        )
+        dao.db.session.flush()
+
+        dao.db.session.add(
+            DraftOutlineItem(
+                outline_item_bid="draft-activity-outline",
+                shifu_bid="draft-activity-course",
+                title="Fresh Lesson",
+                parent_bid="",
+                position="01",
+                prerequisite_item_bids="",
+                llm="",
+                llm_temperature=Decimal("0"),
+                llm_system_prompt="",
+                ask_enabled_status=5101,
+                ask_llm="",
+                ask_llm_temperature=Decimal("0"),
+                ask_llm_system_prompt="",
+                content="",
+                type=0,
+                hidden=0,
+                deleted=0,
+                created_at=datetime(2026, 5, 13, 10, 0, 0),
+                updated_at=datetime(2026, 5, 13, 10, 0, 0),
+                created_user_bid=owner_bid,
+                updated_user_bid=owner_bid,
+            )
+        )
+        dao.db.session.commit()
+
+        result = get_shifu_draft_list(
+            app,
+            owner_bid,
+            page_index=1,
+            page_size=10,
+            is_favorite=False,
+            archived=False,
+            creator_only=True,
+        )
+
+    assert [item.bid for item in result.data[:2]] == [
+        "draft-activity-course",
+        "draft-activity-reference",
     ]
